[rasp/refectoring] 라즈베리파이 리펙토링 완료 및 불피요한 파일 삭제
